amd-features.inc: align with MEL
MEL is now basing its features in a way that disregard
which _particular image_ is being built and gives this
flexibility to user where they can decide which extra
features should be populated. This is manipulated in the
local.conf using EXTRA_IMAGE_FEATURES.
This aligns the AMD features accordingly and basis all
functionality on EXTRA_IMAGE_FEATURES. This would help
minimize the risk of going off track with feature manipulation
while doing open as well as MEL based builds.

Signed-off-by: Awais Belal <awais_belal@mentor.com>