swupdate/initrdscripts: integrate amd-wdt in GRUB cfg for SWU failsafe
This alters the init-install-efi.sh script by integrating the AMD-WDT
in GRUB cfg of MEL installation to implement SWUpdate Failsafe Update
mechanism.

The AMD Watchdog is started as soon as a menuentry is selected by
calling the setup_wdt function. It sets the WDT timeout of 60s by
default which can be changed as needed.

We also introduce a variable "mel_wdt" which allows user to disable
the WDT in case the user is doing kernel debugging or so and wants
the system to keep running.

If WDT is detected to be fired & SWU was performed, then we fallback
to last working state, else we do a normal boot.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>