v1000,r1000/opencl: install only if INCLUDE_OPENCL is set to "yes"
This introduces var INCLUDE_OPENCL into local.conf to control the
inclusion of OpenCL into the image. If user needs OpenCL, they can
add it by setting INCLUDE_OPENCL = "yes" in the local.conf.

This helps in keeping the image size smaller as OpenCL consumes a
lot of disk space.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>