v1000: upgrade vulkan components to 1.1.92
This moves the vulkan bits for the v1000 platform to
1.1.92. The commit updates all the related components
in one go as the vulkan repositories have been reworked
to maintain things in a different way. This copes with all
the conflicts and dependencies.
The tools and headers components use patch version 0 whereas
the loader uses patch version 1 for a particular bug fix.

Signed-off-by: Awais Belal <awais_belal@mentor.com>