common/swu/gdk-pixbuf: rm failing postinsts delayed til first boot
Package post-install intercept hooks that fail while do_rootfs and
generate a warning similar to this:

 WARNING: development-image-1.0-r0 do_rootfs: The postinstall intercept
 hook 'update_pixbuf_cache' failed ...

are deferred for execution on first boot by marking the package as
"unpacked" instead of "installed" in:

 oe-core/meta/lib/oe/package_manager.py
           |-> run_intercepts()
            |-> _postpone_to_first_boot()
             |-> _handle_intercept_failure()
              |-> .mark_packages("unpacked", registered_pkgs.split())

and the oe-core/meta/lib/oe/rootfs.py later on generates an ERROR as:

 ERROR: development-image-1.0-r0 do_rootfs: The following packages
 could not be configured offline and rootfs is read-only: [gdk-pixbuf]

SWUpdate adds "read-only-rootfs" to the IMAGE_FEATURES, hence the
delayed pkg_postinst_* routines cannot be executed on the first boot
to configure the packages.

Therefore as a work-around, in case of gdk-pixbuf, we remove the
pkg_postinst_<pixbuf-pkg> from all the PIXBUF_PACKAGES added in
oe-core/meta/classes/pixbufcache.bbclass so that update_pixbuf_cache
postinst-intercept never applies to these packages.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>