common/swu/libomxil: rm pkg_postinst_ontarget_* because of RO rootfs
Package post-install scripts that are explicitly set to be executed
ontarget and need to create/modify a file to configure the package
cannot execute on a read-only rootfs.

The oe-core/meta/lib/oe/rootfs.py generates an ERROR while do_rootfs:

 ERROR: development-image-1.0-r0 do_rootfs: The following packages
 could not be configured offline and rootfs is read-only: ['libomxil']

SWUpdate adds "read-only-rootfs" to the IMAGE_FEATURES, hence the
delayed pkg_postinst_ontarget_* routines cannot be executed on the
first boot to configure the packages.

Therefore as a work-around, in case of libomxil, we remove the
pkg_postinst_ontarget_* added in common/recipes-multimedia/libomxil.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>