mpv: fix SRC_URI for waf as the current source server has went down
One of the components that our mpv recipe is trying to fetch (i.e. waf)
is not accessible because the server it is trying to get it from
(i.e. freehackers.org) has went down:

  $:~ » ping freehackers.org                     1 ↵
  PING freehackers.org (82.64.38.244) 56(84) bytes of data.
  ^C
  --- freehackers.org ping statistics ---
  4266 packets transmitted,0 received, 100% packet loss, time 4367172ms

...so bitbake falls back to getting it from MIRRORS and the version of
waf available there is different.... so it gets a checksum mismatch and
the build fails:

  ERROR: mpv-0.28.2-r0 do_fetch: Fetcher failure for URL:
  'http://www.freehackers.org/~tnagy/release/waf-1.9.8;name=waf; \
   downloadfilename=waf;subdir=mpv-0.28.2'. Checksum mismatch!

This fixes the SRC_URI for waf to get it from sources.openembedded.org
so that it works normally.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>