docs/SETUP.md: set ground for next release
This moves the following yocto meta layers to the latest commits on the
dunfell branch for next release:
* poky
* meta-openembedded
* meta-dpdk

This also unties meta-amd from the previous release tag.

This also updates the links of yocto docs references from v2.7.2
to v3.1 as being used at the moment for the next release.

This also updates the build system's dependencies as per the current
yocto version 3.1.

Finally, this does some dir renames.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>