amd-spi: rm incompatible spirom kernel module and spirom-test app
This removes the spirom module and app as they are not compatible with
the spi_amd upstream kernel driver that we use now instead of the IOSDK
driver that spirom module and spirom-test app were compatible with.

We will not be able to upgrade the BIOS from userspace for now, but
an alternate way to upgrade the BIOS will be introduced later on.

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>