linux/ccp: make ccp & ccp_crypto modules as some machines dont use them
The Cryptographic Coprocessor (CCP) is not exposed to the x86 CPU on
the ethanolx platforms i.e. rome and milan. Therefore having this
enabled in the kernel by default results in an error at boot time:

[    6.585600] ccp 0000:43:00.1: enabling device (0000 -> 0002)
[    6.590543] ccp 0000:43:00.1: no command queues available
[    6.595151] ccp 0000:43:00.1: sev enabled
[    6.598167] ccp 0000:43:00.1: psp enabled
[    6.833678] ccp 0000:43:00.1: SEV API:1.23 build:23
[    6.837683] ccp_crypto: Cannot load: there are no available CCPs

Lets compile ccp & ccp_crypto as modules so that they can be loaded
when needed by a machine that supports CCP (such as: e3000), and can
be disabled for a machine that does not support it (such as: ethanolx).

Signed-off-by: Arsalan H. Awan <Arsalan_Awan@mentor.com>